PassedGridNotes now list the date the note was seen, not the date of the actual note.
Also it lists the note contents as well as the difference in date, as seen in the notebox. For example, if you passed a global note  that occured 5 years ago, it would say so in the grid, like it does in the daytracker note box.

This required changing the calendar object member in DayTracker to static, so that may cause issues.